ready for PR, closing issue #64 and #61
